pyproject.toml
[project]
name = "finsight-transactions"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "fastapi-pagination>=0.14.0",
    "pydantic-settings[yaml]>=2.10.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.11",
]
